Implement initial version of the input system for entering new entries
Also, both readers now use auxiliary functions for creating each
instance of Objective|Log. This also provides a way for checking if the
user has commited any errors when entering a new objective or log.

In the current state, the program can detect input errors before
appending a new entry into a csv file. This is done provisionally with
the assert instruction.