Finish a working version of routines.edit_delete()
One of the reasons why it was not working on my WSL setup was that I was
using the python.exe while in a Linux session, which meant that I
couldn't use commands normally.
The function now works as intended.
Maybe, now that I know this, I will edit how the menus work,
implementing fzf there as well to improve them.