Replace creation_date with creation_datetime
The first parameter of both Objective and Log is now datetime datatype
instead of just a date.
This was done to allow more precission.

Additionally, both test .csv files where update. Also the readers and
the tests.

Now, when you add an entry by calling the routines.edit_add() function,
the value for creation_datetime is entered automatically with the date
and time of the device